mysite/requirements.txt
google.generativeai
pathlib
SpeechRecognition
streamlit
soundfile